All buttons now work correctly. Validation messages is shown (for errors and updates/creations).

+++ AdBanners/src/test/java/testtask/banners/BannersControllerTest.java
@@ -70,7 +70,7 @@ class BannersControllerTest {
 
   @Test
   public void testGetSingleBanner() throws Exception {
-    when(bannerService.getBanner(any())).thenReturn(banner);
+    when(bannerService.getBanner((Long) any())).thenReturn(banner);
 
     mockMvc.perform(get("/banners/{id}", 1L)
             .contentType(MediaType.APPLICATION_JSON))
